Added command-line arg to retrieve data

+++ main.py
@@ -41,6 +41,8 @@ def purgeDatabase():
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Process cmd line args.')
+    parser.add_argument('--retrieve', dest='retrieve', action='store_true',
+                        default=False)
     parser.add_argument('--init', dest='init', action='store_true',
                         default=False)
     parser.add_argument('--purge', dest='purge', action='store_true',
@@ -48,15 +50,19 @@ if __name__ == "__main__":
     parser.add_argument('--analyze', dest='analyze', action='store_true',
                         default=False)
     args = parser.parse_args()
-    if args.init and not args.purge and not args.analyze:
+    if args.retrieve and not args.init and not args.purge and not args.analyze:
+        print 'Retrieving data...'
+        #retrieveData()
+        print 'Done.'
+    if args.init and not args.retrieve and not args.purge and not args.analyze:
         print 'Initializing database tables...'
         initDatabase()
         print 'Done.'
-    if args.purge and not args.init and not args.analyze:
+    if args.purge and not args.retrieve and not args.init and not args.analyze:
         print 'Purging database (dropping all tables)...'
         purgeDatabase()
         print 'Done.'
-    if args.analyze and not args.init and not args.purge:
+    if args.analyze and not args.retrieve and not args.init and not args.purge:
         print 'Beginning analysis...'
         #analyze()
         print 'Done.'
